+++ AllaganLib.Shared/Misc/Icons.cs
@@ -56,4 +56,5 @@ public static class Icons
     public const int SeedBagIcon = 60928;
     public const int StellarMissionIcon = 65123;
     public const int TripleTriadIcon = 71101;
+    public const int SwordIcon = 63972;
 }